Cleaned up version of the PR-25 commits, which only adds a meta-package
The two commits for PR-25, 777ff89 and c26faee, are very misleading as
they look like they make a huge amount of change, but they acutally don't.
The additional noise is just poor merge management.

The changes in this commit are the net changes on master that occur when
the original PR-25 branch is merged in (ie the diff from the resulting
merge commit on master, to the parent of the merge on master).

The only change here is that the whitespace-only change to
chainjnttojacsolver.cpp was dropped.