Cleaned up version of the PR-58 and PR-61 commits
These commits only add new functions to the chainfksolverxxx classes,
however, the commits are not clean and go back and forward fixing
stuff.

The changes in this commit are the net changes on master that occur when
both the original PR-58 and PR-61 branches are merged in (ie the diff
from the final merge commit on master, to the parent prior to the first
merge on master).