solvers: Only update internal data when necessary
If a chain is updated in a way that doesn't affect the size of
related structures (e.g. only a segment's frameToTip has been updated),
then forcibly resizing internal continers may drop parameters that
have already been set in a solver.

The fix is to only resize when it is necessary, which is when the
size of the chain (either the number of joints or the number of
segments) has changed. This will preserve any internal data that
is only related to the size of the chain size, and not any parameters
contained within the chain's segments.